requirements.txt
reportlab
svglib
ydata-profiling
imblearn
openpyxl
paho-mqtt